Adds routes directory and breaks out routes into multiple files
