+++ server/routes/donation.js
@@ -1,4 +1,4 @@
-const sequelize = require('../db');
+const sequelize = require('../../db');
 
 const DonationController = (router, connection) => {
   /* Donations - Get All */
@@ -35,17 +35,19 @@ const DonationController = (router, connection) => {
   });
 
   /* Donations - Get All by Firebase UID */
-  router.get('/donations/:uid', (req, res) => {});
+  router.get('/donations/:uid', (req, res) => {
+    // Jordan to implement
+  });
 
   /* Donations - Post */
   router.post('/donations', async (req, res) => {
-    if (!req.user) {
-      return res.status(401).send('Must be signed in to post a donation');
-    }
-    const { uid } = req.user;
+    // Jordan - if not user - send 401
     try {
-      const { donation: donationModel, image: imageModel } = connection.models;
+      const { donation: donationModel, image: imageModel, user: userModel } = connection.models;
       const t = await sequelize.transaction();
+
+      // Jordan - come back to this
+      // const donorId = await userModel
       const {
         location,
         description,
@@ -168,87 +170,4 @@ const DonationController = (router, connection) => {
   });
 };
 
-const UserController = (router, connection) => {
-  router.get('/users/:user_id', async (req, res) => {
-    try {
-      const { user: userModel } = connection.models;
-      const { user_id: id } = req.params;
-      const user = await userModel.findOne({ where: { id } });
-      if (!user) {
-        return res.status(404).send('No matching user found');
-      }
-      return res.status(200).send(user);
-    } catch (e) {
-      console.error(e);
-      return res.status(500).end();
-    }
-  });
-
-  router.post('/users', async (req, res) => {
-    if (!req.user) {
-      return res.status(401).send('Unauthorized');
-    }
-    const { uid } = req.user;
-    try {
-      const { user: userModel } = connection.models;
-      const {
-        isIndividual, username, email
-      } = req.body;
-      const defaultLocation = req.body.defaultLocation || null;
-
-      const newUser = await userModel.create({
-        isIndividual,
-        username,
-        email,
-        googleId: uid,
-        defaultLocation,
-      });
-      if (!newUser) {
-        return res.status(406).send('User could not be created');
-      }
-      return res.status(201).send(newUser);
-    } catch (e) {
-      console.error(e);
-      return res.status(500).end();
-    }
-  });
-
-  router.put('/users/:user_id', async (req, res) => {
-    try {
-      const { user: userModel } = connection.models;
-      const { fixedUser } = req.body;
-      const { user_id: id } = req.params;
-      const t = await sequelize.transaction();
-      await userModel.update(
-        {
-          ...fixedUser,
-        },
-        {
-          where: {
-            id,
-          },
-        },
-        { transaction: t }
-      );
-      await t.commit();
-      return res.status(201).end();
-    } catch (e) {
-      console.error(e);
-      return res.status(500).end();
-    }
-  });
-
-  router.delete('/users/:user_id', async (req, res) => {
-    try {
-      const { user: userModel } = connection.models;
-      const { user_id: id } = req.params;
-      await userModel.destroy({ where: { id } });
-      return res.status(201).end();
-    } catch (e) {
-      console.error(e);
-      return res.status(500).end();
-    }
-  });
-};
-
-module.exports = { DonationController, UserController };
+module.exports = DonationController;
