+++ server/routes/index.js
@@ -1,0 +1,4 @@
+const UserController = require('./user');
+const DonationController = require('./donation');
+
+module.exports = { DonationController, UserController };
