+++ server/routes/middleware.js
@@ -1,5 +1,5 @@
 const admin = require('firebase-admin');
-const serviceAccount = require('../firebase-private-key.json');
+const serviceAccount = require('../../firebase-private-key.json');
 
 admin.initializeApp({
   credential: admin.credential.cert(serviceAccount)
@@ -8,10 +8,8 @@ admin.initializeApp({
 async function decodeIDToken(req, res, next) {
   if (req.headers?.authorization?.startsWith('Bearer ')) {
     const idToken = req.headers.authorization.split('Bearer ')[1];
-
     try {
-      const decodedToken = await admin.auth().verifyIdToken(idToken);
-      req.user = decodedToken;
+      req.user = await admin.auth().verifyIdToken(idToken);
     } catch (err) {
       return res.status(401).send(err);
     }
